bundel all provided datasets in the datasets folder
