--- datasets/ausbeer.csv
@@ -1,212 +1,0 @@
-"Y","date"
-284,1956-01-01
-213,1956-04-01
-227,1956-07-01
-308,1956-10-01
-262,1957-01-01
-228,1957-04-01
-236,1957-07-01
-320,1957-10-01
-272,1958-01-01
-233,1958-04-01
-237,1958-07-01
-313,1958-10-01
-261,1959-01-01
-227,1959-04-01
-250,1959-07-01
-314,1959-10-01
-286,1960-01-01
-227,1960-04-01
-260,1960-07-01
-311,1960-10-01
-295,1961-01-01
-233,1961-04-01
-257,1961-07-01
-339,1961-10-01
-279,1962-01-01
-250,1962-04-01
-270,1962-07-01
-346,1962-10-01
-294,1963-01-01
-255,1963-04-01
-278,1963-07-01
-363,1963-10-01
-313,1964-01-01
-273,1964-04-01
-300,1964-07-01
-370,1964-10-01
-331,1965-01-01
-288,1965-04-01
-306,1965-07-01
-386,1965-10-01
-335,1966-01-01
-288,1966-04-01
-308,1966-07-01
-402,1966-10-01
-353,1967-01-01
-316,1967-04-01
-325,1967-07-01
-405,1967-10-01
-393,1968-01-01
-319,1968-04-01
-327,1968-07-01
-442,1968-10-01
-383,1969-01-01
-332,1969-04-01
-361,1969-07-01
-446,1969-10-01
-387,1970-01-01
-357,1970-04-01
-374,1970-07-01
-466,1970-10-01
-410,1971-01-01
-370,1971-04-01
-379,1971-07-01
-487,1971-10-01
-419,1972-01-01
-378,1972-04-01
-393,1972-07-01
-506,1972-10-01
-458,1973-01-01
-387,1973-04-01
-427,1973-07-01
-565,1973-10-01
-465,1974-01-01
-445,1974-04-01
-450,1974-07-01
-556,1974-10-01
-500,1975-01-01
-452,1975-04-01
-435,1975-07-01
-554,1975-10-01
-510,1976-01-01
-433,1976-04-01
-453,1976-07-01
-548,1976-10-01
-486,1977-01-01
-453,1977-04-01
-457,1977-07-01
-566,1977-10-01
-515,1978-01-01
-464,1978-04-01
-431,1978-07-01
-588,1978-10-01
-503,1979-01-01
-443,1979-04-01
-448,1979-07-01
-555,1979-10-01
-513,1980-01-01
-427,1980-04-01
-473,1980-07-01
-526,1980-10-01
-548,1981-01-01
-440,1981-04-01
-469,1981-07-01
-575,1981-10-01
-493,1982-01-01
-433,1982-04-01
-480,1982-07-01
-576,1982-10-01
-475,1983-01-01
-405,1983-04-01
-435,1983-07-01
-535,1983-10-01
-453,1984-01-01
-430,1984-04-01
-417,1984-07-01
-552,1984-10-01
-464,1985-01-01
-417,1985-04-01
-423,1985-07-01
-554,1985-10-01
-459,1986-01-01
-428,1986-04-01
-429,1986-07-01
-534,1986-10-01
-481,1987-01-01
-416,1987-04-01
-440,1987-07-01
-538,1987-10-01
-474,1988-01-01
-440,1988-04-01
-447,1988-07-01
-598,1988-10-01
-467,1989-01-01
-439,1989-04-01
-446,1989-07-01
-567,1989-10-01
-485,1990-01-01
-441,1990-04-01
-429,1990-07-01
-599,1990-10-01
-464,1991-01-01
-424,1991-04-01
-436,1991-07-01
-574,1991-10-01
-443,1992-01-01
-410,1992-04-01
-420,1992-07-01
-532,1992-10-01
-433,1993-01-01
-421,1993-04-01
-410,1993-07-01
-512,1993-10-01
-449,1994-01-01
-381,1994-04-01
-423,1994-07-01
-531,1994-10-01
-426,1995-01-01
-408,1995-04-01
-416,1995-07-01
-520,1995-10-01
-409,1996-01-01
-398,1996-04-01
-398,1996-07-01
-507,1996-10-01
-432,1997-01-01
-398,1997-04-01
-406,1997-07-01
-526,1997-10-01
-428,1998-01-01
-397,1998-04-01
-403,1998-07-01
-517,1998-10-01
-435,1999-01-01
-383,1999-04-01
-424,1999-07-01
-521,1999-10-01
-421,2000-01-01
-402,2000-04-01
-414,2000-07-01
-500,2000-10-01
-451,2001-01-01
-380,2001-04-01
-416,2001-07-01
-492,2001-10-01
-428,2002-01-01
-408,2002-04-01
-406,2002-07-01
-506,2002-10-01
-435,2003-01-01
-380,2003-04-01
-421,2003-07-01
-490,2003-10-01
-435,2004-01-01
-390,2004-04-01
-412,2004-07-01
-454,2004-10-01
-416,2005-01-01
-403,2005-04-01
-408,2005-07-01
-482,2005-10-01
-438,2006-01-01
-386,2006-04-01
-405,2006-07-01
-491,2006-10-01
-427,2007-01-01
-383,2007-04-01
-394,2007-07-01
-473,2007-10-01
-420,2008-01-01
-390,2008-04-01
-410,2008-07-01
