+++ tot/__init__.py
@@ -1,8 +1,13 @@
 import logging
 
 # make version number accessible
-from ._version import __version__  # noqa: F401
-from .benchmark import CrossValidationBenchmark, ManualBenchmark, ManualCVBenchmark, SimpleBenchmark  # noqa: F401
+from ._version import __version__  # noqa: F401 to evade flake8
+from .benchmark import (  # noqa: F401 to evade flake8
+    CrossValidationBenchmark,
+    ManualBenchmark,
+    ManualCVBenchmark,
+    SimpleBenchmark,
+)
 
 # make classes available upon package import
 from .experiment import SimpleExperiment  # noqa: F401
