+++ tot/datasets/dataset.py
@@ -1,6 +1,6 @@
 import logging
 from dataclasses import dataclass, field
-from typing import List, Optional, Tuple, Type
+from typing import List, Optional
 
 import pandas as pd
